edit profile

+++ backend/01-blog/src/main/java/com/__blog/service/posts/PostService.java
@@ -160,10 +160,12 @@ public class PostService {
                 .status(true).data(allPosts).build();
     }
 
-    public ApiResponse<List<PostResponse>> deletePost(UUID postId, UUID userId) {
-        postRepository.deleteById(postId);
-        ApiResponse<List<PostResponse>> allPosts = getPosts(userId);
-        return ApiResponse.<List<PostResponse>>builder()
-                .status(true).data(allPosts.getData()).build();
+    public UUID deletePost(UUID postId) {
+
+        var post = postRepository.findById(postId);
+        if (post.isPresent()) {
+            postRepository.deleteById(postId);
+        }
+        return post.get().getId(); 
     }
 }
